[app] 5 - Moves components to core/comps or features/comps; Renames/adds interfaces/constants-tt-core/feat.ts Updates routes with new import paths;
