+++ src/app/app.component.html
@@ -1,2 +1,1 @@
-<p>app-component works!</p>
 <router-outlet />
