+++ src/app/core/common/interfaces-tt-core.ts
