+++ src/app/features/comps/images-view/images-view.component.ts
