+++ src/app/features/comps/journal-view/journal-view.component.html
