+++ src/app/features/comps/skills-view/skills-view.component.ts
