+++ src/app/features/comps/trades-view/trades-view.component.ts
